Update userGuide.md to match screen layout behaviour (#18112)
Closes #16182

Summary of the issue:
The user guide describes screen layout OFF as "page elements will be
placed on their own lines." This is not entirely correct. Interactive
elements such as buttons and links are placed on a new line. However,
following non-interactive text is left on the same line. This allows the
interactive element to be activated by moving focus to it and pressing
enter. AND it allows easier reading by including non-interactive text on
the same line, thus reducing the number of times down arrow is pressed
to read through text.

Description of user facing changes
Make the description for what happens when Use Screen Layout is off
clearer.

Description of development approach
Edited the user guide entry

Testing strategy:
Text proof read by self and others

Known issues with pull request:
Nil.
Some of the comments in the original issue would also like a change in
behaviour. This PR will make the user guide consistent with the current
behaviour. If a change in behaviour is still required, a new PR can be
created in the future to alter the behaviour, with a corresponding
change in the feature description in the user guide.