Fix the toggle report CLDR script" (#18126)
Follow-up to #18102

Summary of the issue:
When fixing the `GlobalCommands.script_reportCLDR" method, I forgot that you can't just mutate lists in the config, or configobj doesn't know they've changed, so the changes will not be persisted to disc.

Description of user facing changes
The toggle CLDR script will now be persisted across sessions.

Description of development approach
Make a copy of config.conf["speech"]["symbolDictionaries"], mutate it, then save it back.
Initially tried just mutating it and then setting it to itself, but this did not work. Presumably there is some check for identity somewhere, and since all empty lists are identicle, __setitem__ was not called.