Remove Add-on Store panel in secure mode #18129
Closes #17974

Summary of the issue:
The Add-on Store is not available in secure mode and add-ons are not updated in this mode.
Keeping the Add-on Store panel visible with modifiable controls in secure mode is confusing for users.

Moreover, being allow to test an Add-on Store mirror from the logon screen allows to send any GET request from the logon screen, what make these requests not traceable and even maybe insecure.

Description of user facing changes
The add-on store panel is not visible anymore in NVDA's settings window in secure mode.

Description of development approach
Same approach as advanced settings panel.