Fix to case where alt text for math in PDF contains `&` and `<` (#18135)
### Link to issue number:
Fixes #18133 

### Summary of the issue:
In https://github.com/nvaccess/nvda/pull/17276, support was added for
alt text in a formula (a fallback when MathML is not detected) by
wrapping the alt text inside of MathML's mtext tag. That text might have
`<` and `&` characters which should be "escaped" but were not.

The PDF support was added to 2025.1 beta.

The fix is trivial: `<` -> `&lt;` and `&` -> `&amp;` (in that order).

In #18133 I said that the punctuation was not being spoken, but it turns
out that MathCAT actually does convert the punctuation in strings, so
LaTeX is spoken sensibly (but it is still hard to understand when read
quickly).

### Description of user facing changes
In Adobe Acrobat, NVDA now properly reads alt text inside of Formula
tags when no MathML is present (in various forms).

### Description of development approach
Simple string `replace()` calls to do the substitutions.

### Testing strategy:
I opened the test file
[alt.pdf](https://github.com/user-attachments/files/20325821/alt.pdf)
and listened what it said. I checked that in the speech viewer.

### Known issues with pull request:
None

### Code Review Checklist:

- [ ] Documentation:
  - Change log entry
  - User Documentation
  - Developer / Technical Documentation
  - Context sensitive help for GUI changes
- [x] Testing:
  - Unit tests
  - System (end to end) tests
  - Manual testing
- [x] UX of all users considered:
  - Speech
  - Braille
  - Low Vision
  - Different web browsers
  - Localization in other languages / culture than English
- [x] API is compatible with existing add-ons.
- [x] Security precautions taken.

<!-- Please keep the following -->
@coderabbitai summary

---------

Co-authored-by: Sascha Cowley <16543535+SaschaCowley@users.noreply.github.com>
Co-authored-by: Sean Budd <seanbudd123@gmail.com>