fix(build): fix m4 invocation failure when path contains spaces (#18147)
Summary of the issue:
When building NVDA's liblouis component, the build system invokes m4.exe using a path that may include spaces (e.g., in My Codes). However, the path was not wrapped in quotes, causing the command to fail when executed via shell. This resulted in build errors on systems where the path to m4.exe includes spaces.

Description of user facing changes
Description of development approach
Updated the varible env["M4"] in nvdaHelper\liblouis\sconscript to ensure that the path to m4.exe is properly quoted when passed to the shell.