Bugfix: missing tab in spanish symbols. (#18162)
Summary of the issue:
A tab is missing on a symbol in the spanish locale, sending a wrong text to the current speech synthesizer.

Description of user facing changes
As simple as try to read:

≫

With any Spanish voice. The output is questionable:

mucho mayor anone

none is the part of the column