Don't trigger a different configuration profile when Elements list or find dialogs are opened (#18199)
Fixes issue #18160.
Summary of the issue:

When the Elements list dialog, or the Find dialog are opened, NVDA may change the configuration profile, instead of using the current profile, for example, a profile created for the web browser. This behavior doesn't match other NVDA's dialogs, where profile triggers are suspended.
Description of user facing changes

NVDA Will use the current profile when Elements list or Find dialogs are opened.
Description of development approach

shouldSuspendConfigProfileTriggers is set to True for Elements list and Find dialogs.