Move braille table definitions to its own file (#18194)
Related to #11298
Summary of the issue:

Adding new liblouis tables is a manual process for every update. Furthermore, there are tables that are simply missing as they were never added. The display names we use also diverge from the liblouis metadata, which is from a later date.
Description of user facing changes

None
Description of development approach

This pr is the first in a series that will aid in table metadata/addition automation. It moves the add statements for all the tables to a separate file. This file can/should eventually be (re)generated by automation or a script.