Improve server connection failure message (#18223)
Relates to #18196

Summary of the issue:
During the server outage, NVDA users were met with the message "Make
sure you are connected to the internet and try again." when attempting
to check for updates or view the add-on store.
This message is misleading when there is a server outage.
It should be more generic - e.g. cannot connect to the server

Description of user facing changes:
Update wording for UX dialogs for error messages when checking for
updates or fetching data from the add-on store.

Description of developer facing changes:
None

Description of development approach:
String changes