MS Word with UIA: only fetch expandCollapseState for headings   on Word version 16.0.18226 or higher. (#18280)
Fixes #18279
Summary of the issue:

PR #17499 added support for reporting if a heading is collapsed in MS word.
However, this crashes some versions of MS Word before this feature was supported.
For example, MS word version16.0.17932.20396 supports custom attribute values, but crashes if given this specific expandCollapseState constant.
NVDA should limit fetching this data to when it was officially supported in MS Word.
Description of user facing changes:

Microsoft Word 2024 will no longer crash when NVDA tries to fetch the collapsed state of headings.
Description of developer facing changes:
Description of development approach:

Limit fetching of expandcollapseState to MS Word 16.0.18226 and higher.