Fix for the slovenian characters description file (#18304)
Summary of the issue:

The file was not able to be parsed correctly by NVDA, so fixed it.
Description of user facing changes:

All of the slovenian spelling character descriptions should now work.
Description of developer facing changes:

Replaced one space to tab.
Description of development approach:

Replaced space to tab,