Add NVDA logo to installer and uninstaller (#18308)
Summary of the issue:

The installer and uninstaller use the default NSIS logos, which are bad
for branding. Logos are a useful visual cue for identifying the right
program in menus and searches.

Description of user facing changes:
NVDA `N` icon is used for installer and uninstaller.

Description of developer facing changes:
None

Description of development approach:
Added referents to icons in installer and uninstaller script

Testing strategy:
Performing `scons dist` and checking if the files were built with the
right logos.
Performing `scons launcher` and ensuring the installer has the right
logo. Installed the launcher and ensured NVDA used the logo for the
uninstaller.

Known issues with pull request:
None