Voice Access: remove code comment on installed copy requirement. Re #16862.
Voice Access app module will not work with portable NVDA due to lack of admin rights (UI Access) by default. While the source code comment on instlaled copy requirement is removed, the portable copy restriction will remain in the user guide (so that Voice Access can be used as an example of an app with portable copy restrictions).