fix up features for github actions (#18243)
Part of #17878
Summary of the issue:

GitHub actions CI/CD uses feature flags to disable parts of the build.

exit 0 was not cancelling builds if features were disabled.

Additionally, the reasoning for a "symbols" feature isn't clear. Other feature flags are for aspects of the build which integrate with NV Access owned APIs such as Crowdin, Code signing and uploading symbols. Building symbols has no negative side affects other than a slight time increase of the build. Disabling symbols with github actions should be as easy as an if statement in the job.

When system tests are cancelled, the warning message is not correctly posted to the github step summary.
Description of developer facing changes:

When system tests are cancelled, the warning message is now correctly posted to the github step summary.

the symbols feature was removed.

feature_crowdinSync was fixed to check on each step rather than using exit 0