NLS eReader Zoomax braille driver (#18332)
Closes #15863
Copy of #18260
Summary of the issue:

This is the NVDA driver for the NLS eReader Zoomax braille display.
It supports both USB and Bluetooth automatic detection.
Description of user facing changes:

With this driver the user can use directly the NLS eReader Zoomax display without the need to manually install it as an addon.
Description of developer facing changes:
Description of development approach:

The driver is similar with the existing braille display drivers for NVDA.