Add Pyright to GitHub Actions CI (#18345)
Closes #17794

Summary of the issue:
Currently, pyright is not running as part of Pre-commit CI, GitHub
Actions CI, or Appveyor.

Description of user facing changes:
None.

Description of developer facing changes:
Pyright static type analysis now happens as part of CI/CD.

There is now only one pyright action as part of pre-commit, and it is
skipped on pre-commit CI.

Description of development approach:
* Create a power shell script to run pyright and append its output to
the GitHub step summary on failure.

  Add a job to `testAndPublish.yml` that runs this script.

Note the job relies on build NVDA as COM Interfaces may be used as part
of type analysis, but are built by SCons.

* Update type annotations in the SAPI4 driver to fix a Pyright error
* Add `tool.pyright.ignore` to `pyproject.toml` with just `.venv` in it,
to prevent Pyright flagging typing errors in dependencies.

 Testing strategy:
Check workflows on PR.

Known issues with pull request:
None