Fix braille word wrap (#18219)
Fixes #18016
Summary of the issue:

When braille Word wrap is enabled, NVDA doesn't check if the first character of the next braille window is a space, causing that in some cases NVDA Splits text between braille Windows instead of using all braille cells.
Description of user facing changes:

When possible, all braille cells Will be used even if word wrap option is enabled.
Description of developer facing changes:

None.
Description of development approach:

In source/braille.py, the _calculateWindowRowBufferOffsets() function has been modified to check if the next character is a space when Word wrap is True, so that all cells can be used in such cases.