Fix root certificate issues for add-on store (#18354)
Summary of the issue:

If a device doesn't trust NV Access's TLS certificate, fetches to the add-on store fails.
When performing an update check, we update windows root certificates if our certificate is invalid or out of date.
Description of user facing changes:

Users without a trusted NV Access TLS certificate installed already should be able to access the add-on store
Description of developer facing changes:

Code related to updating root certificates is deprecated as it is not intended for the public
Description of development approach:

Made the code to update certificates more generic