Fix for issue #18406 - NLS eReader Zoomax driver does not work with all units (#18414)
Resolves #18406
Summary of the issue:

NLS eReader Zoomax driver does not work with all units.
The issue was detected only in the 2025.2beta1, as the NLS eReader Zoomax driver has now support for autodetection. Previously, the versions of the NLS eReader Zoomax driver were provided as an add-on file, and the autodetection feature was not implemented in that versions.