Make the report language command consistent across settings (#18418)
Fixes #18408
Summary of the issue:

The report language command Works in a different way deppending on certain NVDA's settings. This makes the command les useful. In contrast, it's desirable that the command Works in a consistent way, so that users can change NVDA's settings deppending on the result of this command. For example, they may enable automatic language switching if the language of the text is supported by the current synthesizer, or if it is set correctly in a webpage or document.
Description of user facing changes:

The command to report the current language at caretworks consistently regardless of NVDA's settings.
Description of developer facing changes:

N/A
Description of development approach:

In the globalCommands module, function to get language at caret and script to report it have been updated, removing condition which made them to be based on NVDA's settings.