Only automatically assign milestone on merge of PRs (#18396)
Summary of the issue:
NV Access has a workflow set up to automatically apply a milestone to
closed issues and merged pull requests. While this script only applies a
milestone to issues closed as completed, many issues get closed as
completed even when not planned or duplicates. This results in
unnecessary work for staff.

 Description of user facing changes:
None

Description of developer facing changes:
Only merged PRs are automatically given a milestone.

This matches our approach of applying milestones manually to issues
planned for inclusion in particular releases.

Description of development approach:
Updated `.github/workflows/apply-milestone-on-close.yml` to remove the
code for applying milestones to issues.