Excel: for selection covering more than one cell, announce the active cell when pressing Enter or Tab rather than the whole selection range (#17733)
Closes #6959
Summary of the issue:

When multiple cells are selected, pressing commands that move the active cell (even in a selection) do not allow to report which cell is active in the selection. We should more clearly distinguish between the selection and the active cell in Excel.

This had already been done for Libre Office in #12849.
Description of user facing changes

When multiple cells are selected in an Excel sheet, pressing enter, tab combined or not with shift to move the active cell now reports the new active cell rather than the whole selection.

If needed, the whole selection can still be reported with NVDA+shift+upArrow.
Description of development approach

    Spit the original script in two separate script to handle selection and focus (active cell) reporting separately.
    For charts, the selection is reported in any case as before, there is no concept of active cell