Add WASAPI toggle to SAPI5 (#18352)
Closes #18309.

Summary of the issue:
Users cannot turn off WASAPI for SAPI5 yet.

Description of user facing changes:
Add a synthesizer setting called `useWasapi`, which is displayed as a
checkbox labeled `Use modern audio output system (WASAPI)` in the voice
settings when the built-in SAPI5 synthesizer is selected.

Description of developer facing changes:
None

Description of development approach:
A new `SynthSetting` based on `BooleanDriverSetting` called `useWasapi`
is added and defaults to `True`. It is not in the settings ring.

Legacy SAPI5 code are re-introduced to be used when WASAPI for SAPI5 is
disabled.

Testing strategy:
Tested manually.