Fix SAPI5 (#18473)
Fixup of #18352.
Summary of the issue:

When WASAPI is turned off, bookmark events are not handled properly.
Description of user facing changes:

Users might not notice that, because although the bookmarks are not handled in time, they are still processed at the end of the utterance.
Description of developer facing changes:

None
Description of development approach:

Code that handles the bookmark events when WASAPI is off is added.