Use truststore directly not pip_system_certs (#18490)
Fixes #18460 
Fixup of #18354 

Summary of the issue:
There is a fragile workaround introduced in
https://github.com/nvaccess/nvda/pull/18402#discussion_r2196581863.
This is also a soft API breakage for add-ons, forcing teleNVDA to use
the workaround

This is likely a bug fixed in truststore
https://github.com/sethmlarson/truststore/pull/149

We are using pip_system_certs to patch in an old version of truststore.
this [is no longer being actively
maintained](https://gitlab.com/alelec/pip-system-certs/-/issues/27#note_2372267321)
and apparently truststore has the same patching feature now.

Description of user facing changes:
None

Description of developer facing changes:
Fix workaround which was required for add-ons such as tele NVDA

Testing strategy:
Ensure https://github.com/nvaccess/nvda/issues/18460 is not
reintroduced.
Make sure NVDA can still use update check and the add-on store in
corporate network environments.