Open links in `ui.browseableMessage` in the default browser (#18505)
Summary of the issue:

In an HTML dialog opened by ui.browseableMessage, when a user clicks on a hyperlink, instead of using the system's default browser, a new Internet Explorer window will open and navigate to this link.

This is because ui.browseableMessage uses ShowHTMLDialogEx in mshtml.dll to open an HTML dialog, which is based on Internet Explorer's component. As a result, when a link is being opened, Internet Explorer will be launched.

As we know, Internet Explorer has been retired, and although it may still be fine for simple HTML pages, it is no longer suitable for modern websites.
Description of user facing changes:

Links in the HTML dialog will now open in the system's default browser.