Replace nvda_dmp with an integrated implementation (#18480)
Closes #16633
Summary of the issue:

Right now, NVDA uses nvda_dmp as a proxy to interact with the fast-diff-match-patch library. This approach is inefficient, inflexible, and sometimes unreliable.
Description of how this pull request fixes the issue:

Remove nvda_dmp completely, replacing it with a refactored diffHandler.DiffMatchPatch class that calls DMP directly.