Fixed bug #18386: if there is more than one math expression on a line, everything after the first expression is not spoken/brailled. This includes text along with math.
This has been reported in a few contexts:
* MathCAT: https://github.com/NSoiffer/MathCATForPython/issues/44
* DAISY: https://github.com/daisy/math-a11y/issues/24

The problem lies in code that tries to delete Word's speech. Both the MathML and Word's speech are exposed. The code to recognize this and delete Word's speech was buggy. In particular, it seemed to assume that there was a most one equation on a line. The deletion of Word's math was outside of the loop, so basically everything after the start was deleted. The `controlEnd` command actually pointed to the last MathML on the line, but the speech code only spoke the `controlStart` part, so that bug never showed up.

The fix cleans up the logic and deletes Word's math when it comes to the matching end of the opening MathML.