en Fix bug where only the first math expression on a line was being spoken/brailled #18526
This fixes #18386
Summary of the issue:

If there is more than one math expression on a line, only the first math expression is spoken/brailled. Any text or math after that is dropped.
Description of user facing changes:

Now the full line is read/brailled.
Description of developer facing changes:

None
Description of development approach:

The bug was in wordDocument.py. There was/is special code to avoid saying both the MathML and word's speech for the math. However that code seemed to assume that there was only one piece of math per line and did a single (large) deletion after looping through what was on the line.

The code now correctly deletes only Word's math speech string. It does this in the loop. Because deletion is done on the array as we are looping, the code has to be careful to make sure the index into the array tracks where we are after the deletion.