Added ability to sort add-ons by last tested version (#18506)
Fixes #18440
Summary of the issue:

Sorting add-ons by last tested version can be helpful to know if an add-on is compatible with an updated version of NVDA, available but not installed yet. In this way, users can decide if they prefer to update NVDA, or wait until a compatible version of the add-on is available.
Description of user facing changes:

In add-on lists of the add-on store, a new column with last tested NVDA version is available, and add-ons can also be sorted by this column.
Description of developer facing changes:

None.
Description of development approach:

In addonStoreGui, viewModels, addonList module, a new field with its corresponding text has been added to create a new column, showing the last tested NVDA version